// web/src/app/models/editUserShortInfo.ts
import {User} from "./user";
import {FormGroup} from "@angular/forms";
import {Role} from "./role";

export interface EditUserShortInfo {
  currentData?: UserShortInfo;
  originalData: UserShortInfo;
  editable: boolean;
  validator: FormGroup;
}

export interface UserShortInfo {
  userId?: string;
  login: string;
  name: string;
  surname: string;
  email: string;
  roles: Role[];
}